requirements.txt
crewai>=0.11.0
openai>=1.3.0
langchain>=0.1.0
sqlite3-api>=2.0.1
python-dotenv>=1.0.0 